reorganize server files into sub folders
