--- client/src/AboutFarm.scss
@@ -1,6 +1,0 @@
-@import './styles/variables.scss';
-@import './styles/reset.scss';
-
-.farm {
-  font-size: 150px;
-}
