--- client/src/Account.scss
@@ -1,2 +1,0 @@
-@import './styles/variables.scss';
-@import './styles/reset.scss';
