--- client/src/Home.scss
@@ -1,3 +1,0 @@
-.home {
-  font-size: 200px;
-}
