+++ client/src/components/About/About.scss
@@ -1,5 +1,5 @@
-@import './styles/variables.scss';
-@import './styles/reset.scss';
+@import '../../styles/variables.scss';
+@import '../../styles/reset.scss';
 
 .about__article {
   display:flex;
