+++ client/src/components/AboutFarm/AboutFarm.scss
@@ -1,0 +1,6 @@
+@import '../../styles/variables.scss';
+@import '../../styles/reset.scss';
+
+.farm {
+  font-size: 150px;
+}
