+++ client/src/components/Account/Account.scss
@@ -1,0 +1,2 @@
+@import '../../styles/variables.scss';
+@import '../../styles/reset.scss';
