+++ client/src/components/App.scss
@@ -1,10 +1,10 @@
-@import './styles/variables.scss';
-@import './styles/reset.scss';
+@import '../styles/variables.scss';
+@import '../styles/reset.scss';
 @import "~bootstrap/scss/bootstrap";
 
 body {
   background-color: #7d736a;
-  background-image: url('../public/images/background.jpg');
+  background-image: url('../../public/images/background.jpg');
   background-size: 100%;
   height:100%;
   color: $body-text;
