+++ client/src/components/ContactUs/ContactUs.scss
@@ -1,0 +1,2 @@
+@import '../../styles/variables.scss';
+@import '../../styles/reset.scss';
