+++ client/src/components/HomePage/Home.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import NavButton from './components/home-page/NavButton';
-import BoxCarousel from './components/home-page/BoxCarousel';
+import NavButton from './HomePageComponents/NavButton';
+import BoxCarousel from './HomePageComponents/BoxCarousel';
 import './Home.scss';
 const Home = () => {
   return (
