+++ client/src/components/HomePage/Home.scss
@@ -1,0 +1,6 @@
+@import '../../styles/variables.scss';
+@import '../../styles/reset.scss';
+
+.home {
+  font-size: 200px;
+}
