+++ client/src/components/HomePage/HomePageComponents/BoxCarousel.scss
@@ -1,5 +1,5 @@
-@import "../../styles/mixins.scss";
-@import "../../styles/variables.scss";
+@import "../../../styles/mixins.scss";
+@import "../../../styles/variables.scss";
 
 .div__carousel{
   width: 100%;
