+++ client/src/components/HomePage/HomePageComponents/NavButton.scss
@@ -1,0 +1,6 @@
+@import "../../../styles/mixins.scss";
+@import "../../../styles/variables.scss";
+
+.nav__button {
+  color: blue;
+}
