+++ client/src/components/Navbar/Navbar.scss
@@ -1,3 +1,6 @@
+@import '../../styles/variables.scss';
+@import '../../styles/reset.scss';
+
 nav {
   display: flex;
   justify-content: space-between;
