--- client/src/components/home-page/NavButton.scss
@@ -1,3 +1,0 @@
-.nav__button {
-  color: blue;
-}
